apply pr suggestions
