+++ crates/client/src/schedules.rs
@@ -1,4 +1,6 @@
-use crate::{NamespacedClient, grpc::WorkflowService};
+use crate::{Client, NamespacedClient, grpc::WorkflowService};
+use futures_util::stream;
+use std::collections::VecDeque;
 use std::pin::Pin;
 use std::task::{Context, Poll};
 use std::time::{Duration, SystemTime};
@@ -452,11 +454,10 @@ impl ScheduleDescription {
     /// Convert this description into a [`ScheduleUpdate`] for use with
     /// [`ScheduleHandle::send_update()`].
     ///
-    /// Extracts the schedule definition and conflict token.
+    /// Extracts the schedule definition from the description.
     pub fn into_update(self) -> ScheduleUpdate {
         ScheduleUpdate {
             schedule: self.raw.schedule.unwrap_or_default(),
-            conflict_token: self.raw.conflict_token,
         }
     }
 }
@@ -524,7 +525,6 @@ pub struct ScheduleBackfill {
 #[derive(Debug, Clone)]
 pub struct ScheduleUpdate {
     schedule: schedule_proto::Schedule,
-    conflict_token: Vec<u8>,
 }
 
 impl ScheduleUpdate {
@@ -603,11 +603,6 @@ impl ScheduleUpdate {
     /// Consume and return the raw schedule proto.
     pub fn into_raw(self) -> schedule_proto::Schedule {
         self.schedule
-    }
-
-    /// The conflict token for optimistic concurrency.
-    pub fn conflict_token(&self) -> &[u8] {
-        &self.conflict_token
     }
 
     fn state_mut(&mut self) -> &mut schedule_proto::ScheduleState {
@@ -769,9 +764,8 @@ where
     ///     u.set_note("updated").set_paused(true);
     /// }).await?;
     /// ```
-    // TODO: Add a retry loop for conflict token mismatch once there is a
-    // reliable way to distinguish conflict errors from other failures. Other
-    // SDKs (Ruby, TypeScript) have the same limitation today.
+    // TODO: Add a retry loop for conflict token mismatch. The server
+    // returns FailedPrecondition with "mismatched conflict token".
     pub async fn update(
         &self,
         updater: impl FnOnce(&mut ScheduleUpdate),
@@ -794,7 +788,6 @@ where
                 namespace: self.namespace.clone(),
                 schedule_id: self.schedule_id.clone(),
                 schedule: Some(update.schedule),
-                conflict_token: update.conflict_token,
                 identity: self.client.identity(),
                 request_id: Uuid::new_v4().to_string(),
                 ..Default::default()
@@ -921,5 +914,754 @@ where
         )
         .await?;
         Ok(())
+    }
+}
+
+// Schedule operations on Client.
+impl Client {
+    /// Create a schedule and return a handle to it.
+    pub async fn create_schedule(
+        &self,
+        schedule_id: impl Into<String>,
+        opts: CreateScheduleOptions,
+    ) -> Result<ScheduleHandle<Self>, ScheduleError> {
+        let schedule_id = schedule_id.into();
+        let namespace = self.namespace();
+
+        let initial_patch = if opts.trigger_immediately {
+            Some(schedule_proto::SchedulePatch {
+                trigger_immediately: Some(schedule_proto::TriggerImmediatelyRequest {
+                    // Always use AllowAll for the initial trigger so the
+                    // schedule fires immediately regardless of overlap state.
+                    overlap_policy: ScheduleOverlapPolicy::AllowAll.to_proto(),
+                    scheduled_time: None,
+                }),
+                ..Default::default()
+            })
+        } else {
+            None
+        };
+        // Only send explicit policies when the user set a non-default overlap
+        // policy, so the server uses its own defaults otherwise.
+        let policies = (opts.overlap_policy != ScheduleOverlapPolicy::Unspecified).then(|| {
+            schedule_proto::SchedulePolicies {
+                overlap_policy: opts.overlap_policy.to_proto(),
+                ..Default::default()
+            }
+        });
+        let schedule = schedule_proto::Schedule {
+            spec: Some(opts.spec.into_proto()),
+            action: Some(opts.action.into_proto()),
+            policies,
+            state: Some(schedule_proto::ScheduleState {
+                paused: opts.paused,
+                notes: opts.note,
+                ..Default::default()
+            }),
+        };
+        WorkflowService::create_schedule(
+            &mut self.clone(),
+            CreateScheduleRequest {
+                namespace: namespace.clone(),
+                schedule_id: schedule_id.clone(),
+                schedule: Some(schedule),
+                initial_patch,
+                identity: self.identity(),
+                request_id: Uuid::new_v4().to_string(),
+                ..Default::default()
+            }
+            .into_request(),
+        )
+        .await?;
+        Ok(ScheduleHandle::new(self.clone(), namespace, schedule_id))
+    }
+
+    /// Get a handle to an existing schedule by ID.
+    pub fn get_schedule_handle(&self, schedule_id: impl Into<String>) -> ScheduleHandle<Self> {
+        ScheduleHandle::new(self.clone(), self.namespace(), schedule_id.into())
+    }
+
+    /// List schedules matching the query, returning a stream that lazily
+    /// paginates through results.
+    pub fn list_schedules(&self, opts: ListSchedulesOptions) -> ListSchedulesStream {
+        let client = self.clone();
+        let namespace = self.namespace();
+        let query = opts.query;
+        let page_size = opts.maximum_page_size;
+
+        let stream = stream::unfold(
+            (Vec::new(), VecDeque::new(), false),
+            move |(next_page_token, mut buffer, exhausted)| {
+                let mut client = client.clone();
+                let namespace = namespace.clone();
+                let query = query.clone();
+
+                async move {
+                    if let Some(item) = buffer.pop_front() {
+                        return Some((Ok(item), (next_page_token, buffer, exhausted)));
+                    } else if exhausted {
+                        return None;
+                    }
+
+                    let response = WorkflowService::list_schedules(
+                        &mut client,
+                        ListSchedulesRequest {
+                            namespace,
+                            maximum_page_size: page_size,
+                            next_page_token: next_page_token.clone(),
+                            query,
+                        }
+                        .into_request(),
+                    )
+                    .await;
+
+                    match response {
+                        Ok(resp) => {
+                            let resp = resp.into_inner();
+                            let new_exhausted = resp.next_page_token.is_empty();
+                            let new_token = resp.next_page_token;
+
+                            buffer = resp
+                                .schedules
+                                .into_iter()
+                                .map(ScheduleSummary::from)
+                                .collect();
+
+                            buffer
+                                .pop_front()
+                                .map(|item| (Ok(item), (new_token, buffer, new_exhausted)))
+                        }
+                        Err(e) => Some((Err(e.into()), (next_page_token, buffer, true))),
+                    }
+                }
+            },
+        );
+
+        ListSchedulesStream::new(Box::pin(stream))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::NamespacedClient;
+    use crate::grpc::WorkflowService;
+    use futures_util::FutureExt;
+    use std::sync::Arc;
+    use std::sync::atomic::{AtomicUsize, Ordering};
+    use std::time::SystemTime;
+    use temporalio_common::protos::temporal::api::{
+        common::v1::{
+            Memo, SearchAttributes, WorkflowExecution as ProtoWorkflowExecution, WorkflowType,
+        },
+        schedule::v1::{
+            Schedule, ScheduleActionResult, ScheduleInfo, ScheduleListEntry, ScheduleListInfo,
+            ScheduleSpec, ScheduleState,
+        },
+        workflowservice::v1::{
+            DeleteScheduleResponse, DescribeScheduleResponse, PatchScheduleResponse,
+            UpdateScheduleResponse,
+        },
+    };
+    use tonic::{Request, Response};
+
+    #[derive(Default)]
+    struct CapturedRequests {
+        describe: AtomicUsize,
+        update: AtomicUsize,
+        delete: AtomicUsize,
+        patch: AtomicUsize,
+    }
+
+    #[derive(Clone, Default)]
+    struct MockScheduleClient {
+        captured: Arc<CapturedRequests>,
+        describe_response: DescribeScheduleResponse,
+        should_error: bool,
+    }
+
+    impl NamespacedClient for MockScheduleClient {
+        fn namespace(&self) -> String {
+            "test-namespace".to_string()
+        }
+        fn identity(&self) -> String {
+            "test-identity".to_string()
+        }
+    }
+
+    impl WorkflowService for MockScheduleClient {
+        fn describe_schedule(
+            &mut self,
+            _request: Request<DescribeScheduleRequest>,
+        ) -> futures_util::future::BoxFuture<
+            '_,
+            Result<Response<DescribeScheduleResponse>, tonic::Status>,
+        > {
+            self.captured.describe.fetch_add(1, Ordering::SeqCst);
+            let resp = self.describe_response.clone();
+            let should_error = self.should_error;
+            async move {
+                if should_error {
+                    Err(tonic::Status::not_found("schedule not found"))
+                } else {
+                    Ok(Response::new(resp))
+                }
+            }
+            .boxed()
+        }
+
+        fn update_schedule(
+            &mut self,
+            _request: Request<UpdateScheduleRequest>,
+        ) -> futures_util::future::BoxFuture<
+            '_,
+            Result<Response<UpdateScheduleResponse>, tonic::Status>,
+        > {
+            self.captured.update.fetch_add(1, Ordering::SeqCst);
+            let should_error = self.should_error;
+            async move {
+                if should_error {
+                    Err(tonic::Status::internal("update failed"))
+                } else {
+                    Ok(Response::new(UpdateScheduleResponse::default()))
+                }
+            }
+            .boxed()
+        }
+
+        fn delete_schedule(
+            &mut self,
+            _request: Request<DeleteScheduleRequest>,
+        ) -> futures_util::future::BoxFuture<
+            '_,
+            Result<Response<DeleteScheduleResponse>, tonic::Status>,
+        > {
+            self.captured.delete.fetch_add(1, Ordering::SeqCst);
+            let should_error = self.should_error;
+            async move {
+                if should_error {
+                    Err(tonic::Status::internal("delete failed"))
+                } else {
+                    Ok(Response::new(DeleteScheduleResponse::default()))
+                }
+            }
+            .boxed()
+        }
+
+        fn patch_schedule(
+            &mut self,
+            _request: Request<PatchScheduleRequest>,
+        ) -> futures_util::future::BoxFuture<
+            '_,
+            Result<Response<PatchScheduleResponse>, tonic::Status>,
+        > {
+            self.captured.patch.fetch_add(1, Ordering::SeqCst);
+            let should_error = self.should_error;
+            async move {
+                if should_error {
+                    Err(tonic::Status::internal("patch failed"))
+                } else {
+                    Ok(Response::new(PatchScheduleResponse::default()))
+                }
+            }
+            .boxed()
+        }
+    }
+
+    fn make_schedule_handle(client: MockScheduleClient) -> ScheduleHandle<MockScheduleClient> {
+        ScheduleHandle::new(
+            client,
+            "test-namespace".to_string(),
+            "test-schedule-id".to_string(),
+        )
+    }
+
+    #[test]
+    fn schedule_handle_exposes_namespace_and_id() {
+        let handle = make_schedule_handle(MockScheduleClient::default());
+        assert_eq!(handle.namespace(), "test-namespace");
+        assert_eq!(handle.schedule_id(), "test-schedule-id");
+    }
+
+    #[tokio::test]
+    async fn schedule_describe_returns_response_fields() {
+        let conflict_token = b"token-123".to_vec();
+
+        let client = MockScheduleClient {
+            describe_response: DescribeScheduleResponse {
+                schedule: Some(Schedule::default()),
+                info: Some(ScheduleInfo::default()),
+                memo: Some(Memo {
+                    fields: Default::default(),
+                }),
+                search_attributes: Some(SearchAttributes {
+                    indexed_fields: Default::default(),
+                }),
+                conflict_token: conflict_token.clone(),
+            },
+            ..Default::default()
+        };
+
+        let handle = make_schedule_handle(client.clone());
+        let desc = handle.describe().await.unwrap();
+
+        assert_eq!(client.captured.describe.load(Ordering::SeqCst), 1);
+        assert!(desc.raw().schedule.is_some());
+        assert!(desc.raw().info.is_some());
+        assert!(desc.raw().memo.is_some());
+        assert!(desc.raw().search_attributes.is_some());
+        assert_eq!(desc.conflict_token(), conflict_token);
+    }
+
+    #[tokio::test]
+    async fn schedule_update_describes_then_sends() {
+        let client = MockScheduleClient::default();
+        let handle = make_schedule_handle(client.clone());
+
+        handle
+            .update(|u| {
+                u.set_note("hi");
+            })
+            .await
+            .unwrap();
+
+        assert_eq!(client.captured.describe.load(Ordering::SeqCst), 1);
+        assert_eq!(client.captured.update.load(Ordering::SeqCst), 1);
+    }
+
+    #[tokio::test]
+    async fn schedule_multiple_updates_each_call_service() {
+        let client = MockScheduleClient::default();
+        let handle = make_schedule_handle(client.clone());
+
+        handle.update(|_| {}).await.unwrap();
+        handle.update(|_| {}).await.unwrap();
+
+        assert_eq!(client.captured.update.load(Ordering::SeqCst), 2);
+    }
+
+    #[tokio::test]
+    async fn schedule_delete_calls_service() {
+        let client = MockScheduleClient::default();
+        let handle = make_schedule_handle(client.clone());
+
+        handle.delete().await.unwrap();
+
+        assert_eq!(client.captured.delete.load(Ordering::SeqCst), 1);
+    }
+
+    #[tokio::test]
+    async fn schedule_pause_calls_patch() {
+        let client = MockScheduleClient::default();
+        let handle = make_schedule_handle(client.clone());
+
+        handle.pause(Some("taking a break")).await.unwrap();
+
+        assert_eq!(client.captured.patch.load(Ordering::SeqCst), 1);
+    }
+
+    #[tokio::test]
+    async fn schedule_pause_with_none_uses_default() {
+        let client = MockScheduleClient::default();
+        let handle = make_schedule_handle(client.clone());
+
+        handle.pause(None::<&str>).await.unwrap();
+
+        assert_eq!(client.captured.patch.load(Ordering::SeqCst), 1);
+    }
+
+    #[tokio::test]
+    async fn schedule_unpause_calls_patch() {
+        let client = MockScheduleClient::default();
+        let handle = make_schedule_handle(client.clone());
+
+        handle.unpause(Some("resuming work")).await.unwrap();
+
+        assert_eq!(client.captured.patch.load(Ordering::SeqCst), 1);
+    }
+
+    #[tokio::test]
+    async fn schedule_trigger_calls_patch() {
+        let client = MockScheduleClient::default();
+        let handle = make_schedule_handle(client.clone());
+
+        handle
+            .trigger(ScheduleOverlapPolicy::Unspecified)
+            .await
+            .unwrap();
+
+        assert_eq!(client.captured.patch.load(Ordering::SeqCst), 1);
+    }
+
+    #[tokio::test]
+    async fn schedule_backfill_calls_patch() {
+        let client = MockScheduleClient::default();
+        let handle = make_schedule_handle(client.clone());
+
+        let now = SystemTime::now();
+        handle
+            .backfill(vec![
+                ScheduleBackfill::new(now, now)
+                    .overlap_policy(ScheduleOverlapPolicy::Skip)
+                    .build(),
+                ScheduleBackfill::new(now, now)
+                    .overlap_policy(ScheduleOverlapPolicy::BufferOne)
+                    .build(),
+            ])
+            .await
+            .unwrap();
+
+        assert_eq!(client.captured.patch.load(Ordering::SeqCst), 1);
+    }
+
+    #[tokio::test]
+    async fn schedule_describe_propagates_rpc_errors() {
+        let client = MockScheduleClient {
+            should_error: true,
+            ..Default::default()
+        };
+        let handle = make_schedule_handle(client);
+
+        let err = handle.describe().await.unwrap_err();
+        assert!(
+            matches!(err, ScheduleError::Rpc(_)),
+            "expected Rpc variant, got: {err:?}"
+        );
+        assert!(err.to_string().contains("schedule not found"));
+    }
+
+    #[tokio::test]
+    async fn schedule_update_propagates_rpc_errors() {
+        let client = MockScheduleClient {
+            should_error: true,
+            ..Default::default()
+        };
+        let handle = make_schedule_handle(client);
+
+        let err = handle.update(|_| {}).await.unwrap_err();
+        assert!(matches!(err, ScheduleError::Rpc(_)));
+    }
+
+    #[tokio::test]
+    async fn schedule_delete_propagates_rpc_errors() {
+        let client = MockScheduleClient {
+            should_error: true,
+            ..Default::default()
+        };
+        let handle = make_schedule_handle(client);
+
+        let err = handle.delete().await.unwrap_err();
+        assert!(matches!(err, ScheduleError::Rpc(_)));
+    }
+
+    #[tokio::test]
+    async fn schedule_patch_operations_propagate_rpc_errors() {
+        let client = MockScheduleClient {
+            should_error: true,
+            ..Default::default()
+        };
+        let handle = make_schedule_handle(client);
+
+        assert!(handle.pause(Some("")).await.is_err());
+        assert!(handle.unpause(Some("")).await.is_err());
+        assert!(handle.trigger(Default::default()).await.is_err());
+        assert!(handle.backfill(vec![]).await.is_err());
+    }
+
+    #[tokio::test]
+    async fn schedule_all_patch_operations_call_service() {
+        let client = MockScheduleClient::default();
+        let handle = make_schedule_handle(client.clone());
+
+        handle.pause(Some("p")).await.unwrap();
+        handle.unpause(Some("u")).await.unwrap();
+        handle.trigger(Default::default()).await.unwrap();
+        handle.backfill(vec![]).await.unwrap();
+
+        assert_eq!(client.captured.patch.load(Ordering::SeqCst), 4);
+    }
+
+    #[tokio::test]
+    async fn schedule_describe_accessors_with_populated_fields() {
+        let client = MockScheduleClient {
+            describe_response: DescribeScheduleResponse {
+                schedule: Some(Schedule {
+                    spec: Some(ScheduleSpec {
+                        timezone_name: "US/Eastern".to_string(),
+                        ..Default::default()
+                    }),
+                    state: Some(ScheduleState {
+                        paused: true,
+                        notes: "maintenance window".to_string(),
+                        ..Default::default()
+                    }),
+                    ..Default::default()
+                }),
+                info: Some(ScheduleInfo {
+                    action_count: 42,
+                    missed_catchup_window: 3,
+                    overlap_skipped: 5,
+                    recent_actions: vec![ScheduleActionResult {
+                        start_workflow_result: Some(ProtoWorkflowExecution {
+                            workflow_id: "ra-wf".to_string(),
+                            run_id: "ra-run".to_string(),
+                        }),
+                        ..Default::default()
+                    }],
+                    running_workflows: vec![ProtoWorkflowExecution {
+                        workflow_id: "wf-1".to_string(),
+                        run_id: "run-1".to_string(),
+                    }],
+                    create_time: Some(prost_types::Timestamp {
+                        seconds: 1_700_000_000,
+                        nanos: 0,
+                    }),
+                    update_time: Some(prost_types::Timestamp {
+                        seconds: 1_700_001_000,
+                        nanos: 0,
+                    }),
+                    future_action_times: vec![prost_types::Timestamp {
+                        seconds: 1_700_002_000,
+                        nanos: 0,
+                    }],
+                    ..Default::default()
+                }),
+                conflict_token: b"tok".to_vec(),
+                ..Default::default()
+            },
+            ..Default::default()
+        };
+
+        let handle = make_schedule_handle(client);
+        let desc = handle.describe().await.unwrap();
+
+        assert!(desc.paused());
+        assert_eq!(desc.note(), Some("maintenance window"));
+        assert_eq!(desc.action_count(), 42);
+        assert_eq!(desc.missed_catchup_window(), 3);
+        assert_eq!(desc.overlap_skipped(), 5);
+
+        assert_eq!(desc.recent_actions().len(), 1);
+        assert_eq!(
+            desc.running_actions(),
+            vec![ScheduleRunningAction {
+                workflow_id: "wf-1".to_string(),
+                run_id: "run-1".to_string(),
+            }]
+        );
+
+        assert_eq!(desc.future_action_times().len(), 1);
+        assert!(desc.create_time().is_some());
+        assert!(desc.update_time().is_some());
+    }
+
+    #[tokio::test]
+    async fn schedule_describe_defaults_when_nested_fields_are_none() {
+        let client = MockScheduleClient::default();
+
+        let handle = make_schedule_handle(client);
+        let desc = handle.describe().await.unwrap();
+
+        assert!(!desc.paused());
+        assert_eq!(desc.note(), None);
+        assert_eq!(desc.action_count(), 0);
+        assert_eq!(desc.missed_catchup_window(), 0);
+        assert_eq!(desc.overlap_skipped(), 0);
+        assert!(desc.recent_actions().is_empty());
+        assert!(desc.running_actions().is_empty());
+        assert!(desc.future_action_times().is_empty());
+        assert!(desc.create_time().is_none());
+        assert!(desc.update_time().is_none());
+        assert!(desc.conflict_token().is_empty());
+    }
+
+    #[tokio::test]
+    async fn schedule_note_returns_none_for_empty_string() {
+        let client = MockScheduleClient {
+            describe_response: DescribeScheduleResponse {
+                schedule: Some(Schedule {
+                    state: Some(ScheduleState {
+                        notes: String::new(),
+                        ..Default::default()
+                    }),
+                    ..Default::default()
+                }),
+                ..Default::default()
+            },
+            ..Default::default()
+        };
+
+        let handle = make_schedule_handle(client);
+        let desc = handle.describe().await.unwrap();
+        assert_eq!(desc.note(), None);
+    }
+
+    #[test]
+    fn schedule_summary_note_returns_none_for_empty_string() {
+        let entry = ScheduleListEntry {
+            schedule_id: "s".to_string(),
+            info: Some(ScheduleListInfo {
+                notes: String::new(),
+                ..Default::default()
+            }),
+            ..Default::default()
+        };
+        let summary = ScheduleSummary::from(entry);
+        assert_eq!(summary.note(), None);
+    }
+
+    #[test]
+    fn schedule_summary_accessors() {
+        let entry = ScheduleListEntry {
+            schedule_id: "sched-1".to_string(),
+            memo: Some(Memo {
+                fields: Default::default(),
+            }),
+            search_attributes: Some(SearchAttributes {
+                indexed_fields: Default::default(),
+            }),
+            info: Some(ScheduleListInfo {
+                spec: Some(ScheduleSpec::default()),
+                workflow_type: Some(WorkflowType {
+                    name: "MyWorkflow".to_string(),
+                }),
+                notes: "some note".to_string(),
+                paused: true,
+                recent_actions: vec![ScheduleActionResult {
+                    start_workflow_result: Some(ProtoWorkflowExecution {
+                        workflow_id: "ra-wf".to_string(),
+                        run_id: "ra-run".to_string(),
+                    }),
+                    ..Default::default()
+                }],
+                future_action_times: vec![prost_types::Timestamp {
+                    seconds: 1_700_000_000,
+                    nanos: 0,
+                }],
+            }),
+        };
+
+        let summary = ScheduleSummary::from(entry);
+        assert_eq!(summary.schedule_id(), "sched-1");
+        assert!(summary.raw().memo.is_some());
+        assert!(summary.raw().search_attributes.is_some());
+        assert_eq!(summary.workflow_type(), Some("MyWorkflow"));
+        assert_eq!(summary.note(), Some("some note"));
+        assert!(summary.paused());
+        assert_eq!(summary.recent_actions().len(), 1);
+        assert_eq!(summary.future_action_times().len(), 1);
+    }
+
+    #[test]
+    fn schedule_summary_defaults_when_info_is_none() {
+        let entry = ScheduleListEntry {
+            schedule_id: "sched-2".to_string(),
+            ..Default::default()
+        };
+
+        let summary = ScheduleSummary::from(entry);
+        assert_eq!(summary.schedule_id(), "sched-2");
+        assert!(summary.raw().memo.is_none());
+        assert!(summary.raw().search_attributes.is_none());
+        assert_eq!(summary.workflow_type(), None);
+        assert_eq!(summary.note(), None);
+        assert!(!summary.paused());
+        assert!(summary.recent_actions().is_empty());
+        assert!(summary.future_action_times().is_empty());
+    }
+
+    #[test]
+    fn schedule_description_raw_round_trip() {
+        let resp = DescribeScheduleResponse {
+            conflict_token: b"ct".to_vec(),
+            schedule: Some(Schedule::default()),
+            ..Default::default()
+        };
+        let desc = ScheduleDescription::from(resp.clone());
+        assert_eq!(desc.raw().conflict_token, b"ct");
+        let recovered = desc.into_raw();
+        assert_eq!(recovered.conflict_token, resp.conflict_token);
+        assert!(recovered.schedule.is_some());
+    }
+
+    #[test]
+    fn schedule_summary_raw_round_trip() {
+        let entry = ScheduleListEntry {
+            schedule_id: "rt-1".to_string(),
+            ..Default::default()
+        };
+        let summary = ScheduleSummary::from(entry.clone());
+        assert_eq!(summary.raw().schedule_id, "rt-1");
+        let recovered = summary.into_raw();
+        assert_eq!(recovered.schedule_id, entry.schedule_id);
+    }
+
+    #[test]
+    fn schedule_into_update_preserves_schedule() {
+        let resp = DescribeScheduleResponse {
+            schedule: Some(Schedule {
+                state: Some(ScheduleState {
+                    notes: "my notes".to_string(),
+                    ..Default::default()
+                }),
+                ..Default::default()
+            }),
+            ..Default::default()
+        };
+        let desc = ScheduleDescription::from(resp);
+        let update = desc.into_update();
+
+        assert_eq!(update.raw().state.as_ref().unwrap().notes, "my notes");
+    }
+
+    #[test]
+    fn schedule_update_setters_are_chainable() {
+        let resp = DescribeScheduleResponse {
+            schedule: Some(Schedule::default()),
+            ..Default::default()
+        };
+        let desc = ScheduleDescription::from(resp);
+        let mut update = desc.into_update();
+        update.set_note("chained").set_paused(true);
+        assert_eq!(update.raw().state.as_ref().unwrap().notes, "chained");
+        assert!(update.raw().state.as_ref().unwrap().paused);
+    }
+
+    #[test]
+    fn schedule_recent_action_from_proto_with_timestamps() {
+        let ts = prost_types::Timestamp {
+            seconds: 1_700_000_000,
+            nanos: 0,
+        };
+        let proto = ScheduleActionResult {
+            schedule_time: Some(ts),
+            actual_time: Some(ts),
+            start_workflow_result: Some(ProtoWorkflowExecution {
+                workflow_id: "wf-abc".to_string(),
+                run_id: "run-xyz".to_string(),
+            }),
+            ..Default::default()
+        };
+
+        let action = ScheduleRecentAction::from(&proto);
+
+        assert!(action.schedule_time.is_some());
+        assert!(action.actual_time.is_some());
+        assert_eq!(action.workflow_id, "wf-abc");
+        assert_eq!(action.run_id, "run-xyz");
+    }
+
+    #[test]
+    #[should_panic(expected = "unsupported schedule action")]
+    fn schedule_recent_action_panics_without_workflow_result() {
+        let _ = ScheduleRecentAction::from(&ScheduleActionResult::default());
+    }
+
+    #[test]
+    fn schedule_overlap_policy_default_is_unspecified() {
+        assert_eq!(
+            ScheduleOverlapPolicy::default(),
+            ScheduleOverlapPolicy::Unspecified
+        );
     }
 }
